Split test assist components to their own files
Now they have templates associated with them it's cleaner to put them in
their own files, otherwise it's hard to see which templates belong to
which components
